크롤러 수정 및 북마크 생성시 리턴값 생성

+++ database/migrations/2019_08_22_053247_create_bookmarks_table.php
@@ -15,8 +15,8 @@ class CreateBookmarksTable extends Migration
     {
         Schema::create('bookmarks', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('category_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('category_id');
             $table->string('url');
             $table->string('site_name')->nullable();
             $table->string('title')->nullable();
@@ -26,6 +26,7 @@ class CreateBookmarksTable extends Migration
             $table->timestamps();
         });
     }
+
 
     /**
      * Reverse the migrations.
